+++ frontend/src/routes/Module/components/MetadataSidebarBlock.tsx
@@ -1,66 +1,11 @@
-import {
-  MetadataSidebarBlock,
-  MetadataSidebarBlockItem,
-} from "@/components/MetadataSidebarBlock";
-import { github } from "@/icons/github";
-import { document } from "@/icons/document";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { getModuleVersionDataQuery } from "../query";
 import { useModuleParams } from "../hooks/useModuleParams";
-import { definitions } from "@/api";
 
-function getLinkLabel(url: string) {
-  const match = url.match(/github\.com\/([^/]+)\/([^/]+)/);
-
-  if (!match) {
-    return null;
-  }
-
-  return `${match[1]}/${match[2]}`;
-}
-
-interface BlockProps {
-  license?: definitions["LicenseList"];
-  link?: string | undefined;
-}
-
-function Block(props: BlockProps) {
-  return (
-    <MetadataSidebarBlock title="Repository">
-      <MetadataSidebarBlockItem icon={document} title="License">
-        {props.license ? (
-          props.license.map((license) => (
-            <a
-              href={license.link}
-              key={license.spdx}
-              className="underline"
-              target="_blank"
-              rel="noreferrer noopener"
-            >
-              {license.spdx}
-            </a>
-          ))
-        ) : (
-          <span className="flex h-em w-24 animate-pulse bg-gray-500/25" />
-        )}
-      </MetadataSidebarBlockItem>
-      <MetadataSidebarBlockItem icon={github} title="GitHub">
-        {props.link ? (
-          <a
-            href={props.link}
-            className="underline"
-            target="_blank"
-            rel="noreferrer noopener"
-          >
-            {getLinkLabel(props.link)}
-          </a>
-        ) : (
-          <span className="flex h-em w-32 animate-pulse bg-gray-500/25" />
-        )}
-      </MetadataSidebarBlockItem>
-    </MetadataSidebarBlock>
-  );
-}
+import {
+  RepoMetadataSidebarBlock,
+  RepoMetadataSidebarBlockSkeleton,
+} from "@/components/RepoMetadataSidebarBlock";
 
 export function ModuleMetadataSidebarBlock() {
   const { namespace, name, target, version } = useModuleParams();
@@ -69,9 +14,7 @@ export function ModuleMetadataSidebarBlock() {
     getModuleVersionDataQuery(namespace, name, target, version),
   );
 
-  return <Block license={data.licenses} link={data.link} />;
+  return <RepoMetadataSidebarBlock license={data.licenses} link={data.link} />;
 }
 
-export function ModuleMetadataSidebarBlockSkeleton() {
-  return <Block />;
-}
+export { RepoMetadataSidebarBlockSkeleton as ModuleMetadataSidebarBlockSkeleton };
